LPS-153277 Move CodeMirrorEditor to frontend-js-components-web
