+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/code_editor/Editor.scss
