+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/code_editor/Sidebar.scss
@@ -1,7 +1,7 @@
 @import 'atlas-variables';
 
 .lfr-objects__code-editor-sidebar {
-	background-color: $light-l1;
+	background-color: $light-l2;
 	border-left: 1px solid $gray-300;
 	overflow-x: hidden;
 	width: 280px;
@@ -46,9 +46,5 @@
 		display: flex;
 		justify-content: flex-end;
 		width: 20%;
-	}
-
-	&-collapsible-button-list {
-		margin-bottom: 15px;
 	}
 }
