+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/field_base/FieldBase.scss
@@ -1,7 +1,7 @@
 @import 'atlas-variables';
 
-.lfr-objects__field-base {
-	&-tooltip-icon {
+.lfr-field-base {
+	&__tooltip-icon {
 		color: $secondary-l1;
 	}
 }
