+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/field_base/FieldBase.tsx
@@ -70,7 +70,7 @@ export function FieldBase({
 					<ClayTooltipProvider>
 						<span data-tooltip-align="top" title={tooltip}>
 							<ClayIcon
-								className="lfr-objects__field-base-tooltip-icon"
+								className="lfr-field-base__tooltip-icon"
 								symbol="question-circle-full"
 							/>
 						</span>
