+++ modules/apps/frontend-js/frontend-js-components-web/src/main/resources/META-INF/resources/field_base/FieldFeedback.scss
@@ -1,6 +1,6 @@
 @import 'atlas-variables';
 
-.lfr-de__field-feedback {
+.lfr-field-base__feedback {
 	color: $secondary;
 	font-size: 0.875rem;
