+++ modules/apps/frontend-js/frontend-js-components-web/types/src/main/resources/META-INF/resources/index.d.ts
@@ -1,0 +1,44 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+/**
+ * Code Editor
+ */
+export {default as CodeEditor, SidebarCategory} from './code_editor/index';
+export {default as CodeMirrorEditor} from './code_editor/Editor';
+
+/**
+ * Field Base
+ */
+export * from './field_base/index';
+
+/**
+ * Management Toolbar
+ */
+export {default as ManagementToolbar} from './management_toolbar/ManagementToolbar';
+
+/**
+ * Translation Manager
+ */
+export {
+	activeLanguageIdsAtom,
+	selectedLanguageIdAtom,
+} from './translation_manager/state';
+export {default as TranslationAdminModal} from './translation_manager/TranslationAdminModal';
+export {default as TranslationAdminSelector} from './translation_manager/TranslationAdminSelector';
+
+/**
+ * Treeview
+ */
+export {default as Treeview} from './treeview/Treeview';
