+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/components/AutoComplete/index.tsx
@@ -13,9 +13,8 @@
  */
 
 import ClayDropDown from '@clayui/drop-down';
+import {FieldBase} from 'frontend-js-components-web';
 import React, {useState} from 'react';
-
-import {FieldBase} from '../FieldBase';
 
 import './index.scss';
 import {CustomSelect} from './CustomSelect';
