+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/components/Input.tsx
@@ -13,9 +13,8 @@
  */
 
 import {ClayInput} from '@clayui/form';
+import {FieldBase} from 'frontend-js-components-web';
 import React from 'react';
-
-import {FieldBase} from './FieldBase';
 
 export const Input = React.forwardRef<HTMLInputElement, IProps>(
 	(
