+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/components/Select.tsx
@@ -13,9 +13,8 @@
  */
 
 import {ClaySelect} from '@clayui/form';
+import {FieldBase} from 'frontend-js-components-web';
 import React from 'react';
-
-import {FieldBase} from './FieldBase';
 
 interface ISelectProps extends React.SelectHTMLAttributes<HTMLSelectElement> {
 	disabled?: boolean;
