+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/components/SelectWithOption.tsx
@@ -13,9 +13,8 @@
  */
 
 import {ClaySelect, ClaySelectWithOption} from '@clayui/form';
+import {FieldBase} from 'frontend-js-components-web';
 import React from 'react';
-
-import {FieldBase} from './FieldBase';
 
 export function SelectWithOption({
 	ariaLabel,
