+++ modules/apps/object/object-js-components-web/src/main/resources/META-INF/resources/ts_modules/frontend-js-state-web.d.ts
@@ -12,6 +12,4 @@
  * details.
  */
 
-declare module '@liferay/frontend-js-state-web' {
-	export * from '@liferay/frontend-js-state-web/src/main/resources/META-INF/resources/State';
-}
+declare module '@liferay/frontend-js-state-web';
