+++ modules/apps/object/object-js-components-web/types/src/main/resources/META-INF/resources/components/ExpressionBuilder.d.ts
@@ -12,8 +12,8 @@
  * details.
  */
 
+import {SidebarCategory} from 'frontend-js-components-web';
 import React from 'react';
-import {SidebarCategory} from './CodeEditor/index';
 export declare function ExpressionBuilder({
 	buttonDisabled,
 	className,
