+++ modules/apps/object/object-web/package.json
@@ -1,6 +1,5 @@
 {
 	"dependencies": {
-		"@types/codemirror": "^5.60.5",
 		"codemirror": "^5.59.0",
 		"text-mask-core": "^5.1.2"
 	},
