+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/DataValidation/ObjectValidationTabs.d.ts
@@ -13,7 +13,7 @@
  */
 
 import 'codemirror/mode/groovy/groovy';
-import {SidebarCategory} from '@liferay/object-js-components-web';
+import {SidebarCategory} from 'frontend-js-components-web';
 import {ChangeEventHandler} from 'react';
 import {ObjectValidationErrors} from '../ObjectValidationFormBase';
 export declare function BasicInfo({
