+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ObjectAction/tabs/ActionBuilder.d.ts
@@ -14,7 +14,8 @@
 
 /// <reference types="react" />
 
-import {CustomItem, SidebarCategory} from '@liferay/object-js-components-web';
+import {CustomItem} from '@liferay/object-js-components-web';
+import {SidebarCategory} from 'frontend-js-components-web';
 import './ActionBuilder.scss';
 import {ActionError} from '../index';
 export default function ActionBuilder({
